+++ config/support.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Roberts/Support
-return [
-
-];