+++ src/Traits/HasCreator.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Roberts\Support\Traits;
+
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+trait HasCreator
+{
+    protected static function bootHasCreator()
+    {
+        static::creating(function ($model) {
+            if (Auth::check()) {
+                $model->creator_id = Auth::id();
+            }
+        });
+    }
+
+    public function creator(): BelongsTo
+    {
+        return $this->belongsTo(config('auth.providers.users.model'), 'creator_id');
+    }
+}