+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
 All notable changes to `support` will be documented in this file.
+
+## 12.0.0 - 2025-09-11
+
+- support laravel 12
+- add hascreator trait
+- add hasupdater trait
+- add randomorcreate test helper